move `Exception.h` to its own Bazel component
