+++ common/common.cc
@@ -1,5 +1,5 @@
 #include "common/common.h"
-#include "common/Exception.h"
+#include "common/exception/Exception.h"
 #include "common/FileOps.h"
 #include "common/sort.h"
 #include "os/os.h"
