+++ common/common.h
@@ -170,6 +170,6 @@ std::string demangle(const char *mangled);
 #pragma GCC poison cuserid
 #pragma GCC poison rexec rexec_af
 
-#include "Exception.h"
+#include "exception/Exception.h"
 #include "Timer.h"
 #endif
